Add improved logging functionality.
