+++ blinkpy/helpers/util.py
@@ -121,18 +121,3 @@ class BlinkURLHandler():
         self.networks_url = "{}/networks".format(self.base_url)
         self.video_url = "{}/api/v2/videos".format(self.base_url)
         _LOGGER.debug("Setting base url to %s.", self.base_url)
-
-
-class RepeatLogHandler(logging.StreamHandler):
-    """Log handler for repeat entries."""
-
-    def __init__(self):
-        """Initialize repeat log handler."""
-        super().__init__()
-        self.log_record = set()
-
-    def emit(self, record):
-        """Ensure we only log a message once."""
-        if record.msg not in self.log_record:
-            self.log_record.add(record.msg)
-            super().emit(record)
